refactor(web): share URL search-param update helpers
